Add keyboard shortcuts
git-svn-id: file:///.data/svn_convert/svn@460 f7a67266-ba00-4dfb-90a2-3eec4ba07b7d

+++ trunk/sim/msgview.cpp
@@ -557,7 +557,7 @@ int MsgView::findMsg(unsigned long id, int parag)
             unsigned i;
             for (i = 0; i < s.length(); i++)
                 if (s[(int)i] == '.') break;
-            if (atol((const char*)s + i + 1) == id)
+            if ((unsigned long)(atol((const char*)s + i + 1)) == id)
                 return parag;
         }
     }
